Extract the `MarketPageModel` from the `MarketPage` file.
